it's broken but i'm testing?

+++ src/backend/api/express.ts
@@ -6,8 +6,10 @@ import * as bodyParser from 'body-parser'
 import { PORT } from '..'
 import { Auth, authMiddleware } from './endpoints/auth'
 import { User } from './endpoints/user'
+import { Events } from './endpoints/events'
 import { Mongo } from '../db/mongo'
 import { User as MongoUser } from '../db/models/user'
+import { UploadUsage } from '../db/models/upload-usage'
 import * as multer from 'multer'
 import * as fs from 'fs'
 import { HelixUser } from '@twurple/api/lib'
@@ -51,6 +53,7 @@ export class Api {
 
     Auth.bind()
     User.bind()
+    Events.bind()
 
 
     Api.endpoints.get('/api/logger', async (req,res) => {
@@ -113,28 +116,50 @@ export class Api {
     })
 
     Api.endpoints.route('/api/uploads/:userId/:filename')
-      .post(authMiddleware, Api.upload, async (req, res) => {
+      .post(authMiddleware, Api.upload, async (req, res): Promise<any> => {
+        let autoIndexing = req.headers['autoindexing']==='true'
+
         //@ts-ignore
         let file = req.file
         let readStream = Readable.from(file.buffer)
+
+        let plain = req.params.filename.replace(/\.[^.]+$/gi,'')
+        let extension = req.params.filename.replace(plain+'.','')  
+        let i=0
+
         let found: any = await Mongo.Upload.findOne({ filename: req.params.filename, 'metadata.userId': Mongo.ObjectId(req.params.userId) })
-        if(found) await new Promise((resolve, reject) => {
-          Mongo.Upload.unlink({ _id: found._id}, (error, unlink) => {
-            if(error) reject(error)
-            else resolve(unlink)
-          })
+        if(found) await new Promise(async (resolve, reject) => {
+          if(!autoIndexing) {
+            await UploadUsage.remove({ fileId: found._id })
+            await Mongo.Upload.unlink({ _id: found._id}, (error, unlink) => {
+              if(error) reject(error)
+              else resolve(unlink)
+            })
+            
+          } else {
+            while(found) {
+              found = await Mongo.Upload.findOne({ filename: `${plain}-${++i}.${extension}`, 'metadata.userId': Mongo.ObjectId(req.params.userId) })
+            }
+            resolve(found)
+          }
         })
+
         await Mongo.Upload.write({
-          filename: file.originalname,
-          metadata: { userId: Mongo.ObjectId(req.params.userId) },
+          filename: `${plain}${i>0 ? `-${i}` : ''}.${extension}`,
+          metadata: { userId: Mongo.ObjectId(req.params.userId),  },
           contentType: file.mimetype
-        }, readStream, (error, f): any => {
+        }, readStream, async (error, f): Promise<any> => {
           if(error) return res.status(500).send()
+           await UploadUsage.create({
+            fileId: f._id,
+            usages: 1
+          })
           res.send({
             //url: /image\//.test(file.mimetype) ? `uploads/${req.params.userId}/${file.originalname}` : null
-            url: `uploads/${req.params.userId}/${file.originalname}`
+            url: `uploads/${req.params.userId}/${plain}${i>0 ? `-${i}` : ''}.${extension}`
           })
         })
+
       })
       .get(async (req, res): Promise<any> => {
         let found: any = await Mongo.Upload.findOne({ filename: req.params.filename, 'metadata.userId': Mongo.ObjectId(req.params.userId) })
@@ -174,14 +199,42 @@ export class Api {
       .delete(authMiddleware, async (req, res): Promise<any> => {
         let found: any = await Mongo.Upload.findOne({ filename: req.params.filename, metadata: { userId: Mongo.ObjectId(req.params.userId) } })
         if(!found) return res.status(404).send()
-        await new Promise((resolve, reject) => {
+        await new Promise(async (resolve, reject) => {
+          let usage: any = UploadUsage.findOne({fileId: found._id})
+          if(+usage.usages>1) {
+            usage.usages = +usage.usages -1
+            await usage.save()
+          } else 
+            await UploadUsage.remove({ fileId: found._id })
+
           Mongo.Upload.unlink({ _id: found._id}, (error, unlink) => {
             if(error) reject(error)
             else resolve(unlink)
           })
         })
-        res.send()
+        res.send({})
       })
+
+    Api.endpoints.get('/api/uploads/:userId/link/:filename', async (req, res): Promise<any> => {
+      let found: any = await Mongo.Upload.findOne({ filename: req.params.filename, 'metadata.userId': Mongo.ObjectId(req.params.userId) })
+      if(!found) return res.status(404).send()
+      let usages: any = await UploadUsage.findOne({ fileId: found._id }) 
+      if(!usages) {
+          usages = await UploadUsage.create({
+          fileId: found._id,
+          usages: 1
+        }) 
+      } else  
+        usages.usages = +usages.usages + 1
+
+      await usages.save()
+      res.send({})
+    })
+
+    Api.endpoints.get('/api/uploads/:userId/unlink/:filename', async (req, res): Promise<any> => {
+      await Api.unlink(req.params.filename, req.params.userId, res)
+    })
+    
 
 
 /*     Api.endpoints.get('/api/status', async (req, res) => {
@@ -194,5 +247,27 @@ export class Api {
 
 
   
+  }
+
+  static async unlink (filename, userId, res?): Promise<any> {
+    let found: any = await Mongo.Upload.findOne({ filename: filename, 'metadata.userId': Mongo.ObjectId(userId) })
+    if(!found) return res ? res.status(404).send() : null
+    let usages: any = await UploadUsage.findOne({ fileId: found._id }) 
+    if(!usages) return res ? res.send(`Nothing to do!`) : null
+    
+    if(+usages.usages<=1) {
+      await UploadUsage.deleteOne({ _id: usages._id})
+      await new Promise(async (resolve, reject) => {
+        Mongo.Upload.unlink({ _id: found._id}, (error, unlink) => {
+          if(error) reject(error)
+          else resolve(unlink)
+        })
+      })
+    } else {
+      usages.usages = +usages.usages - 1
+      await usages.save()
+    }
+    if(res) res.send({})
+    return 
   }
 }
